changes amount type

+++ account-service/src/main/java/com/digitalmoneyhouse/accountservice/dto/TransactionRequest.java
@@ -1,5 +1,7 @@
 package com.digitalmoneyhouse.accountservice.dto;
 
+import java.math.BigDecimal;
+
 import com.digitalmoneyhouse.accountservice.model.TransactionType;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
@@ -14,7 +16,7 @@ public class TransactionRequest {
 
     @NotNull
     @Positive
-    private Double amount;
+    private BigDecimal amount;
 
     @NotNull
     private TransactionType type;
@@ -25,26 +27,26 @@ public class TransactionRequest {
 
     private String destinationAccount;
 
-    public TransactionRequest(Double amount, TransactionType type, String destinationAccount) {
+    public TransactionRequest(BigDecimal amount, TransactionType type, String destinationAccount) {
         this.amount = amount;
         this.type = type;
         this.destinationAccount = destinationAccount;
     }
 
-    public TransactionRequest(Double amount, TransactionType type, String description, String destinationAccount) {
+    public TransactionRequest(BigDecimal amount, TransactionType type, String description, String destinationAccount) {
         this.amount = amount;
         this.type = type;
         this.description = description;
         this.destinationAccount = destinationAccount;
     }
 
-    public TransactionRequest(Double amount, TransactionType type, Integer cardId) {
+    public TransactionRequest(BigDecimal amount, TransactionType type, Integer cardId) {
         this.amount = amount;
         this.type = type;
         this.cardId = cardId;
     }
 
-    public TransactionRequest(Double amount, TransactionType type, String description, Integer cardId) {
+    public TransactionRequest(BigDecimal amount, TransactionType type, String description, Integer cardId) {
         this.amount = amount;
         this.type = type;
         this.description = description;
